Made the download for GDELT work.
Now skips local files (zips of csvs).

+++ src/GdeltDownloader.py
@@ -1,3 +1,6 @@
+#!/usr/bin/env python3
+
+from src.timer import timer
 from typing import Dict, List, Tuple, Optional, Union, Iterator
 from tqdm import tqdm
 from bs4 import BeautifulSoup as bs
@@ -14,12 +17,13 @@ import re
 import sys
 import shutil
 import logging
-logging.basicConfig(level=logging.INFO)
+
+logging.basicConfig(level=logging.WARNING)
 
 
 class GdeltDownloader(object):
 
-    def __init__(self, url: str, start_date: Tuple[int, ...], end_date: Tuple[int, ...]):
+    def __init__(self, url: str, start_date: Tuple[int, int, int], end_date: Tuple[int, int, int]):
         now = datetime.now()
         start_date if start_date else (1970, 1, 1)
         end_date if end_date else (now.year, now.month, now.day)
@@ -42,17 +46,15 @@ class GdeltDownloader(object):
 
         file_links = []
         response = requests.get(self.base_url)
-        # print(response.content)
 
         soup = bs(response.text, features="lxml")
 
         links = soup.find_all('li')
-        # print(links)
 
         for link in tqdm(links, desc="Parsing links ..."):
             try:
                 file = link.find('a')['href']
-                date_string = tuple(int(x)
+                date_string: Tuple[int, int, int] = tuple(int(x)
                                     for x in re.findall(dateRegex, file)[0])
                 date = datetime(*date_string)
 
@@ -67,12 +69,13 @@ class GdeltDownloader(object):
                     file_links.append(link_obj)
 
             except Exception as e:
-                # print("Link was empty!", e)
+                logging.info(f"Link '{link}' was empty!", e)
                 continue  # Skip link if one of the properties was not found
 
-        print(f"{len(file_links)} links found!")
+        print(f"{len(file_links)} links found in date range '{self.start_date} - {self.end_date}'!")
         return file_links
 
+    #TODO: Change return type to a dict.
     def download_file(self, file_obj: Dict, dl_path: str, extract: bool = True, remove: bool = True) -> Optional[Tuple[str, bool]]:
         """Downloads and extract a given file. Skips the file if it already exists.
 
@@ -88,15 +91,21 @@ class GdeltDownloader(object):
         Optional[Tuple[str, bool]]
             Return a tuple with information about the download. First item is the local filename, second item is a bool which is set to true if the md5 sum of the file matches. Can be none if the download was skipped.
         """
-        local_filename = file_obj['file'].split('/')[-1]
-        file_path = os.path.join(dl_path, local_filename)
+        local_filename = file_obj['file'].split('/')[-1]  # Get filename from file_obj for later use.
+        file_path = os.path.join(dl_path, local_filename)  # Compose relative file path. This is where the file gets saved.
 
         result = None
 
-        logging.info(f"{file_path}, exists: {os.path.isfile(file_path)}")
+        try:
+            size = os.path.getsize(file_path)
+            if size == 0:
+                os.remove(file_path)
+        except os.error as e:
+            logging.debug(f"File '{file_path}' does not exist.")
 
-        if not os.path.isfile(file_path):
-            with requests.get(file_path, stream=True) as r:
+        # Skip download if .zip or .csv file already exists.
+        if not os.path.isfile(file_path) and not os.path.isfile(os.path.splitext(file_path)[0]):
+            with requests.get(os.path.join(self.base_url, local_filename), stream=True) as r:
                 with open(file_path, "wb") as f:
                     shutil.copyfileobj(r.raw, f)
                     is_md5_equal = (
@@ -104,13 +113,16 @@ class GdeltDownloader(object):
 
             result = (local_filename, is_md5_equal)
         else:
-            print(f"Skipping download for: {file_path}")
+            logging.debug(f"Skipping download for: {file_path}.")
 
-        if extract:
+        if extract and not os.path.isfile(os.path.splitext(file_path)[0]):
             self.unzip(file_path, dl_path)
+        else:
+            logging.debug(f"Skipping extraction of {file_path}.")
 
         return result
 
+    @timer
     def download_links(self, files: List[Dict], dl_path: str, thread_count: int = 1) -> Iterator[Optional[Tuple[str, bool]]]:
         """Takes a dict and dispatches the download links to different threads to download.
 
@@ -129,7 +141,6 @@ class GdeltDownloader(object):
             Iterator object with download results.
         """
 
-        logging.info(files[0:2])
         _ = input("Press any key to start download and extraction process ...")
         executor = ThreadPoolExecutor(max_workers=thread_count)
         results = tqdm(executor.map(
@@ -153,6 +164,7 @@ class GdeltDownloader(object):
         with open(file_path, "rb") as f:
             for chunk in iter(lambda: f.read(4096), b""):
                 md5().update(chunk)
+
         return md5().hexdigest()
 
     def unzip(self, file_path: str, extract_path: str, remove: bool = True) -> None:
@@ -165,11 +177,13 @@ class GdeltDownloader(object):
         extract_path : str
             Destination path for extracted file.
         """
+        logging.debug(f"Unpacking file '{file_path}' to '{extract_path}'")
+
         with ZipFile(file_path, mode="r") as zipobj:
             zipobj.extractall(path=extract_path)
 
         if remove:
-            os.remove(file_path)  # Remove file after done unzipping
+            os.remove(file_path)  # Remove file after unzipping is done.
 
 
 if __name__ == "__main__":
